requirements.txt
anyio==3.6.1
app-store-scraper==0.3.5
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
cssselect==1.1.0
DateTime==4.7
fake-useragent==0.1.11
fastapi==0.85.0
feedparser==6.0.10
google-play-scraper==1.2.2
gunicorn==20.1.0
h11==0.14.0
idna==2.10
importlib-metadata==4.12.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
parse==1.19.0
pydantic==1.10.2
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
requests==2.23.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.25.11
uvicorn==0.18.3
w3lib==2.0.1
websockets==10.3
zipp==3.8.1
zope.interface==5.4.0